+++ src/modules/platform-tools/hb-service/hb-service.service.ts
@@ -1,7 +1,8 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, GoneException, BadRequestException } from '@nestjs/common';
 import { ConfigService } from '../../../core/config/config.service';
+import { Logger } from '../../../core/logger/logger.service';
 
 @Injectable()
 export class HbServiceService {
@@ -9,6 +10,7 @@ export class HbServiceService {
 
   constructor(
     private readonly configService: ConfigService,
+    private readonly logger: Logger
   ) { }
 
   /**
@@ -58,10 +60,30 @@ export class HbServiceService {
     return fs.writeJsonSync(this.hbServiceSettingsPath, settings, { spaces: 4 });
   }
 
+  /**
+   * Set the flag to trigger a full restart on next boot
+   */
   async setFullServiceRestartFlag() {
     // restart ui on next restart
     this.configService.hbServiceUiRestartRequired = true;
 
     return { status: 0 };
+  }
+
+  /**
+   * Stream the full log file to the client
+   */
+  async downloadLogFile() {
+    if (!await fs.pathExists(this.configService.ui.log.path)) {
+      this.logger.error(`Cannot download log file: "${this.configService.ui.log.path}" does not exist.`);
+      throw new BadRequestException(`Log file not found on disk.`);
+    }
+    try {
+      await fs.access(this.configService.ui.log.path, fs.constants.R_OK);
+    } catch (e) {
+      this.logger.error(`Cannot download log file: Missing read permissions on "${this.configService.ui.log.path}".`);
+      throw new BadRequestException('Cannot read log file. Check the log file permissions');
+    }
+    return fs.createReadStream(this.configService.ui.log.path, 'utf8');
   }
 }